feat(list): ajust component architecture and layout , add icons
