+++ package.json
@@ -23,6 +23,7 @@
     "react-native-reanimated": "~4.1.1",
     "react-native-safe-area-context": "~5.6.0",
     "react-native-screens": "~4.16.0",
+    "react-native-svg": "^15.13.0",
     "react-native-worklets": "0.5.1"
   },
   "private": true,
